// frontend/app/interface.ts
export interface IEnhancedStock {
  name: string;
  ticker: string;
  quantity: number;
  buy_price: number;
  current_price: number;
  profit_loss_percent: number;
}